+++ openrem/remapp/interface/chart_functions.py
@@ -30,6 +30,7 @@
 
 """
 
+import os
 import math
 import base64
 from builtins import range  # pylint: disable=redefined-builtin
@@ -48,10 +49,10 @@ from scipy import stats
 
 
 def global_config(
-        filename,
-        height_multiplier=1.0,
-        height=1080,
-        width=1920,
+    filename,
+    height_multiplier=1.0,
+    height=1080,
+    width=1920,
 ):
     """
     Creates a Plotly global configuration dictionary. The parameters all relate
@@ -80,6 +81,7 @@ def create_dataframe(
     database_events,
     field_dict,
     data_point_name_lowercase=None,
+    data_point_name_remove_whitespace_padding=None,
     data_point_value_multipliers=None,
     uid=None,
 ):
@@ -92,6 +94,7 @@ def create_dataframe(
     :param field_dict: a dictionary of lists, each containing database field names to include in the DataFrame. The
                        dictionary should include "names", "values", "dates", "times" and optionally "system" items
     :param data_point_name_lowercase: boolean flag to determine whether to make all "names" field values lower case
+    :param data_point_name_remove_whitespace_padding: boolean flag to determine whether to strip whitespace
     :param data_point_value_multipliers: list of float valuse to multiply each "values" field value by
     :param uid: string containing database field name which contains a unique identifier for each record
     :return: a Pandas DataFrame with a column per required field
@@ -110,41 +113,69 @@ def create_dataframe(
     fields_to_include.update(field_dict["times"])
     fields_to_include.update(field_dict["system"])
 
+    start = None
+    if settings.DEBUG:
+        start = datetime.now()
+
     # NOTE: I am not excluding zero-value events from the calculations (zero DLP or zero CTDI)
+
+    # The "order_by()" in the command below removes the custom ordering on the query set that is used to order things
+    # correctly on the filtered page tables. This ordering isn't required for the DataFrame; removing it speeds up
+    # the DataFrame.from_records command.
     df = pd.DataFrame.from_records(
-        data=database_events.values_list(*fields_to_include),  # values_list uses less memory than values
+        data=database_events.order_by().values_list(
+            *fields_to_include
+        ),  # values_list uses less memory than values
         columns=fields_to_include,  # need to specify the column names as we're now using values_list
         coerce_float=True,  # force Decimal to float - saves doing a type conversion later
     )
 
-    dtype_conversion = {}
-    for name_field in field_dict["names"]:
-        dtype_conversion[name_field] = "category"
+    if settings.DEBUG:
+        print(f"Initial Dataframe created from records in {datetime.now() - start}")
+        start = datetime.now()
+        print("Initial DataFrame info, including memory use, is:")
+        df.info()
 
-        # Replace any empty values with "Blank" (Plotly doesn't like empty values)
-        df[name_field].fillna(value="Blank", inplace=True)
-        # Make lowercase if required
-        if data_point_name_lowercase:
-            df[name_field] = df[name_field].str.lower()
+    # Replace any NaN values in the names columns with "Blank"
+    df[field_dict["names"]] = df[field_dict["names"]].apply(lambda x: x.fillna("Blank"))
 
+    # Make names column values lowercase if required
+    if data_point_name_lowercase:
+        df[field_dict["names"]] = df[field_dict["names"]].apply(lambda x: x.str.lower())
+
+    # Strip whitespace from the beginning and end of any names column values
+    # Also replace multiple spaces with a single space
+    if data_point_name_remove_whitespace_padding:
+        df[field_dict["names"]] = df[field_dict["names"]].apply(lambda x: x.str.strip().replace("\s+", " ", regex=True))
+
+    # Make the names columns all "category" type - this saves memory. Must be done after the above, as the string
+    # replacement lines revert the columns back to "object"
+    df[field_dict["names"]] = df[field_dict["names"]].astype("category")
+
+    # Rename the "system" column to "x_ray_system_name" if it is present
     if field_dict["system"]:
         df.rename(columns={field_dict["system"][0]: "x_ray_system_name"}, inplace=True)
+        df["x_ray_system_name"] = df["x_ray_system_name"].astype("category")
         df.sort_values(by="x_ray_system_name", inplace=True)
+    # Else create the "x_ray_system_name" column populated with a single "All systems" category
     else:
-        df.insert(0, "x_ray_system_name", "All systems")
-    dtype_conversion["x_ray_system_name"] = "category"
+        df["x_ray_system_name"] = pd.Categorical(np.full(len(df.index), "All systems"))
 
+    # Loop through each value field, multiplying the values by the corresponding multiplier
     for idx, value_field in enumerate(field_dict["values"]):
         if data_point_value_multipliers:
             df[value_field] *= data_point_value_multipliers[idx]
 
+    # Convert each date field to a pd datetime using a specific date format
     for date_field in field_dict["dates"]:
         df[date_field] = pd.to_datetime(df[date_field], format="%Y-%m-%d")
 
-    df = df.astype(dtype_conversion)
-
     if settings.DEBUG:
-        print(f"Dataframe created in {datetime.now() - start}")
+        print(
+            f"Dataframe fillna, lower case, whitespace stripping etc took {datetime.now() - start}"
+        )
+        print("DataFrame info after processing, including memory use, is:")
+        df.info()
 
     return df
 
@@ -258,7 +289,7 @@ def plotly_set_default_theme(theme_name):
     pio.templates.default = theme_name
 
 
-def calculate_colour_sequence(scale_name="jet", n_colours=10):
+def calculate_colour_sequence(scale_name="RdYlBu", n_colours=10):
     """
     Calculates a sequence of n_colours from the matplotlib colourmap scale_name
 
@@ -325,23 +356,42 @@ def csv_data_barchart(fig, params):
     fig_data_dict = fig.to_dict()["data"]
 
     if params["df_name_col"] != "performing_physician_name":
-        df = pd.DataFrame(data=fig_data_dict[0]["x"], columns=[params["name_axis_title"]])
+        df = pd.DataFrame(
+            data=fig_data_dict[0]["x"], columns=[params["name_axis_title"]]
+        )
         for data_set in fig_data_dict:
             new_col_df = pd.DataFrame(
                 data=data_set["customdata"][:, 1:],
-                columns=[data_set["name"] + " " + params["value_axis_title"], "Frequency"]
+                columns=[
+                    data_set["name"] + " " + params["value_axis_title"],
+                    "Frequency",
+                ],
             )
             df = pd.concat([df, new_col_df], axis=1)
 
         return df
 
     else:
-        df = pd.DataFrame(data=fig_data_dict[0]["x"], columns=[params["name_axis_title"]])
+        df = pd.DataFrame(
+            data=fig_data_dict[0]["x"], columns=[params["name_axis_title"]]
+        )
         for data_set in fig_data_dict:
-            series_name = data_set["hovertemplate"].split(params["facet_col"] + "=")[1].split("<br>")[0]
-            new_col_df = pd.DataFrame(data=data_set["customdata"][:, 1:],  # pylint: disable=line-too-long
-                                      columns=[data_set["name"] + " " + series_name + " " + params["value_axis_title"], "Frequency"]  # pylint: disable=line-too-long
-                                      )
+            series_name = (
+                data_set["hovertemplate"]
+                .split(params["facet_col"] + "=")[1]
+                .split("<br>")[0]
+            )
+            new_col_df = pd.DataFrame(
+                data=data_set["customdata"][:, 1:],  # pylint: disable=line-too-long
+                columns=[
+                    data_set["name"]
+                    + " "
+                    + series_name
+                    + " "
+                    + params["value_axis_title"],
+                    "Frequency",
+                ],  # pylint: disable=line-too-long
+            )
             df = pd.concat([df, new_col_df], axis=1)
         return df
 
@@ -358,7 +408,9 @@ def csv_data_frequency(fig, params):
 
     df = pd.DataFrame(data=fig_data_dict[0]["x"], columns=[params["x_axis_title"]])
     for data_set in fig_data_dict:
-        df = pd.concat([df, pd.DataFrame(data=data_set["y"], columns=[data_set["name"]])], axis=1)
+        df = pd.concat(
+            [df, pd.DataFrame(data=data_set["y"], columns=[data_set["name"]])], axis=1
+        )
 
     return df
 
@@ -378,6 +430,37 @@ def calc_facet_rows_and_height(df, facet_col_name, facet_col_wrap):
     if chart_height < 500:
         chart_height = 500
     return chart_height, n_facet_rows
+
+
+def save_fig_as_html_div(fig, filename, active=settings.SAVE_CHARTS_AS_HTML):
+    """
+    Saves the Plotly figure as an HTML file containing a single DIV. The file is saved on the OpenREM server in
+    MEDIA_ROOT\charts\yyyy\mm\dd\. Viewing the saved file requires an active internet connection as the Plotly
+    JavaScript library is not included in the file.
+
+    This method is not currently accessible to an OpenREM user or administrator - it is present to assist developers
+    when producing example charts for the OpenREM documentation. It must be manually activated by setting active=True
+    in the method definition.
+
+    Args:
+        fig: a Plotly figure
+        filename: (string )the filename to use
+        active: (boolean) to set whether to save the figure
+    """
+    if active:
+        datestamp = datetime.now()
+        path = os.path.join(
+            settings.MEDIA_ROOT, "charts", datestamp.strftime("%Y/%m/%d")
+        )
+        os.makedirs(
+            os.path.join(settings.MEDIA_ROOT, "charts", datestamp.strftime("%Y/%m/%d")),
+            exist_ok=True,
+        )
+        fig.write_html(
+            os.path.join(path, filename + ".html"),
+            include_plotlyjs="cdn",
+            full_html=False,
+        )
 
 
 def plotly_boxplot(
@@ -411,7 +494,9 @@ def plotly_boxplot(
             return empty_dataframe_msg()
 
         if params["facet_col"]:
-            chart_height, n_facet_rows = calc_facet_rows_and_height(df, params["facet_col"], params["facet_col_wrap"])
+            chart_height, n_facet_rows = calc_facet_rows_and_height(
+                df, params["facet_col"], params["facet_col_wrap"]
+            )
 
         n_colours = len(df.x_ray_system_name.unique())
         colour_sequence = calculate_colour_sequence(params["colourmap"], n_colours)
@@ -437,14 +522,13 @@ def plotly_boxplot(
         fig.update_traces(quartilemethod="exclusive")
 
         fig.update_xaxes(
-            tickson="boundaries",
-            ticks="outside",
-            ticklen=5,
-            showticklabels=True
+            tickson="boundaries", ticks="outside", ticklen=5, showticklabels=True
         )
         fig.update_yaxes(showticklabels=True, matches=None)
 
         fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
+
+        save_fig_as_html_div(fig, params["filename"])
 
         if params["return_as_dict"]:
             return fig.to_dict()
@@ -453,13 +537,15 @@ def plotly_boxplot(
                 fig,
                 output_type="div",
                 include_plotlyjs=False,
-                config=global_config(params["filename"], height_multiplier=chart_height / 500.0),
+                config=global_config(
+                    params["filename"], height_multiplier=chart_height / 500.0
+                ),
             )
 
     except ValueError as e:
         return failed_chart_message_div(
             "Could not resolve chart. Try filtering the data to reduce the number of systems.",
-            e
+            e,
         )
 
 
@@ -560,7 +646,9 @@ def plotly_barchart(
     n_facet_rows = 1
 
     if params["facet_col"]:
-        chart_height, n_facet_rows = calc_facet_rows_and_height(df, params["facet_col"], params["facet_col_wrap"])
+        chart_height, n_facet_rows = calc_facet_rows_and_height(
+            df, params["facet_col"], params["facet_col_wrap"]
+        )
 
     n_colours = len(df.x_ray_system_name.unique())
     colour_sequence = calculate_colour_sequence(params["colourmap"], n_colours)
@@ -592,14 +680,13 @@ def plotly_barchart(
     )
 
     fig.update_xaxes(
-        tickson="boundaries",
-        ticks="outside",
-        ticklen=5,
-        showticklabels=True
+        tickson="boundaries", ticks="outside", ticklen=5, showticklabels=True
     )
     fig.update_yaxes(showticklabels=True, matches=None)
 
     fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
+
+    save_fig_as_html_div(fig, params["filename"])
 
     if params["return_as_dict"]:
         return fig.to_dict(), None
@@ -609,12 +696,17 @@ def plotly_barchart(
             csv_name,
         )
 
-        return plot(
-            fig,
-            output_type="div",
-            include_plotlyjs=False,
-            config=global_config(params["filename"], height_multiplier=chart_height / 500.0),
-        ), csv_data
+        return (
+            plot(
+                fig,
+                output_type="div",
+                include_plotlyjs=False,
+                config=global_config(
+                    params["filename"], height_multiplier=chart_height / 500.0
+                ),
+            ),
+            csv_data,
+        )
 
 
 def plotly_histogram_barchart(
@@ -645,7 +737,9 @@ def plotly_histogram_barchart(
     # pylint: disable=too-many-locals
     # pylint: disable=too-many-branches
     # pylint: disable=too-many-statements
-    chart_height, n_facet_rows = calc_facet_rows_and_height(df, params["df_facet_col"], params["facet_col_wrap"])
+    chart_height, n_facet_rows = calc_facet_rows_and_height(
+        df, params["df_facet_col"], params["facet_col_wrap"]
+    )
 
     n_colours = len(df[params["df_category_col"]].unique())
     colour_sequence = calculate_colour_sequence(params["colourmap"], n_colours)
@@ -654,11 +748,15 @@ def plotly_histogram_barchart(
     mid_bins = None
     bin_labels = None
     if params["global_max_min"]:
-        bin_labels, bins, mid_bins = calc_histogram_bin_data(df, params["df_value_col"], n_bins=params["n_bins"])
+        bin_labels, bins, mid_bins = calc_histogram_bin_data(
+            df, params["df_value_col"], n_bins=params["n_bins"]
+        )
 
     try:
         fig = make_subplots(
-            rows=n_facet_rows, cols=params["facet_col_wrap"], vertical_spacing=0.40 / n_facet_rows
+            rows=n_facet_rows,
+            cols=params["facet_col_wrap"],
+            vertical_spacing=0.40 / n_facet_rows,
         )
 
         current_row = 1
@@ -744,6 +842,8 @@ def plotly_histogram_barchart(
         fig.update_layout(layout)
         fig.update_layout(legend_title_text=params["legend_title"])
 
+        save_fig_as_html_div(fig, params["filename"])
+
         if params["return_as_dict"]:
             return fig.to_dict()
         else:
@@ -751,13 +851,15 @@ def plotly_histogram_barchart(
                 fig,
                 output_type="div",
                 include_plotlyjs=False,
-                config=global_config(params["filename"], height_multiplier=chart_height / 500.0),
+                config=global_config(
+                    params["filename"], height_multiplier=chart_height / 500.0
+                ),
             )
 
     except ValueError as e:
         return failed_chart_message_div(
             "Could not resolve chart. Try filtering the data to reduce the number of categories or systems.",
-            e
+            e,
         )
 
 
@@ -810,14 +912,18 @@ def plotly_binned_statistic_barchart(
     # pylint: disable=too-many-locals
     # pylint: disable=too-many-branches
     # pylint: disable=too-many-statements
-    chart_height, n_facet_rows = calc_facet_rows_and_height(df, params["df_facet_col"], params["facet_col_wrap"])
+    chart_height, n_facet_rows = calc_facet_rows_and_height(
+        df, params["df_facet_col"], params["facet_col_wrap"]
+    )
 
     n_colours = len(df[params["df_category_col"]].unique())
     colour_sequence = calculate_colour_sequence(params["colourmap"], n_colours)
 
     try:
         fig = make_subplots(
-            rows=n_facet_rows, cols=params["facet_col_wrap"], vertical_spacing=0.40 / n_facet_rows
+            rows=n_facet_rows,
+            cols=params["facet_col_wrap"],
+            vertical_spacing=0.40 / n_facet_rows,
         )
 
         current_row = 1
@@ -911,7 +1017,9 @@ def plotly_binned_statistic_barchart(
 
             if current_col == 1:
                 fig.update_yaxes(
-                    title_text=params["stat_name"].capitalize() + " " + params["y_axis_title"],
+                    title_text=params["stat_name"].capitalize()
+                    + " "
+                    + params["y_axis_title"],
                     row=current_row,
                     col=current_col,
                 )
@@ -927,6 +1035,8 @@ def plotly_binned_statistic_barchart(
         fig.update_layout(layout)
         fig.update_layout(legend_title_text=params["facet_title"])
 
+        save_fig_as_html_div(fig, params["filename"])
+
         if params["return_as_dict"]:
             return fig.to_dict()
         else:
@@ -934,13 +1044,15 @@ def plotly_binned_statistic_barchart(
                 fig,
                 output_type="div",
                 include_plotlyjs=False,
-                config=global_config(params["file_name"], height_multiplier=chart_height / 500.0),
+                config=global_config(
+                    params["filename"], height_multiplier=chart_height / 500.0
+                ),
             )
 
     except ValueError as e:
         return failed_chart_message_div(
             "Could not resolve chart. Try filtering the data to reduce the number of categories or systems.",
-            e
+            e,
         )
 
 
@@ -970,7 +1082,9 @@ def plotly_timeseries_linechart(
     :return: Plotly figure embedded in an HTML DIV; or Plotly figure as a dictionary (if "return_as_dict" is True);
              or an error message embedded in an HTML DIV if there was a ValueError when calculating the figure
     """
-    chart_height, n_facet_rows = calc_facet_rows_and_height(df, params["facet_col"], params["facet_col_wrap"])
+    chart_height, n_facet_rows = calc_facet_rows_and_height(
+        df, params["facet_col"], params["facet_col_wrap"]
+    )
 
     n_colours = len(df[params["df_name_col"]].unique())
     colour_sequence = calculate_colour_sequence(params["colourmap"], n_colours)
@@ -1016,6 +1130,8 @@ def plotly_timeseries_linechart(
 
         fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
 
+        save_fig_as_html_div(fig, params["filename"])
+
         if params["return_as_dict"]:
             return fig.to_dict()
         else:
@@ -1023,13 +1139,15 @@ def plotly_timeseries_linechart(
                 fig,
                 output_type="div",
                 include_plotlyjs=False,
-                config=global_config(params["filename"], height_multiplier=chart_height / 500.0),
+                config=global_config(
+                    params["filename"], height_multiplier=chart_height / 500.0
+                ),
             )
 
     except ValueError as e:
         return failed_chart_message_div(
             "Could not resolve chart. Try filtering the data to reduce the number of categories or systems.",
-            e
+            e,
         )
 
 
@@ -1052,12 +1170,14 @@ def plotly_scatter(
     :param params["colourmap"]: (string) colourmap to use
     :param params["x_axis_title"]: (string) x-axis title
     :param params["y_axis_title"]: (string) y-axis title
-    :param params["file_name"]: (string) default filename to use for plot bitmap export
+    :param params["filename"]: (string) default filename to use for plot bitmap export
     :param params["return_as_dict"]: (boolean) flag to trigger return as a dictionary rather than a HTML DIV
     :return: Plotly figure embedded in an HTML DIV; or Plotly figure as a dictionary (if "return_as_dict" is True);
              or an error message embedded in an HTML DIV if there was a ValueError when calculating the figure
     """
-    sorted_category_list = create_sorted_category_list(df, params["df_name_col"], params["df_y_col"], params["sorting"])
+    sorted_category_list = create_sorted_category_list(
+        df, params["df_name_col"], params["df_y_col"], params["sorting"]
+    )
 
     params["df_category_name_col"] = params["df_name_col"]
     params["df_group_col"] = "x_ray_system_name"
@@ -1072,7 +1192,9 @@ def plotly_scatter(
         if df.empty:
             return empty_dataframe_msg()
 
-        chart_height, n_facet_rows = calc_facet_rows_and_height(df, params["df_group_col"], params["facet_col_wrap"])
+        chart_height, n_facet_rows = calc_facet_rows_and_height(
+            df, params["df_group_col"], params["facet_col_wrap"]
+        )
 
         n_colours = len(df[params["df_category_name_col"]].unique())
         colour_sequence = calculate_colour_sequence(params["colourmap"], n_colours)
@@ -1104,6 +1226,8 @@ def plotly_scatter(
 
         fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
 
+        save_fig_as_html_div(fig, params["filename"])
+
         if params["return_as_dict"]:
             return fig.to_dict()
         else:
@@ -1111,13 +1235,15 @@ def plotly_scatter(
                 fig,
                 output_type="div",
                 include_plotlyjs=False,
-                config=global_config(params["file_name"], height_multiplier=chart_height / 500.0),
+                config=global_config(
+                    params["filename"], height_multiplier=chart_height / 500.0
+                ),
             )
 
     except ValueError as e:
         return failed_chart_message_div(
             "Could not resolve chart. Try filtering the data to reduce the number of categories or systems.",
-            e
+            e,
         )
 
 
@@ -1147,7 +1273,9 @@ def plotly_barchart_weekdays(
     :return: Plotly figure embedded in an HTML DIV; or Plotly figure as a dictionary (if "return_as_dict" is True);
              or an error message embedded in an HTML DIV if there was a ValueError when calculating the figure
     """
-    chart_height, n_facet_rows = calc_facet_rows_and_height(df, "x_ray_system_name", facet_col_wrap)
+    chart_height, n_facet_rows = calc_facet_rows_and_height(
+        df, "x_ray_system_name", facet_col_wrap
+    )
 
     try:
         fig = px.bar(
@@ -1191,6 +1319,8 @@ def plotly_barchart_weekdays(
 
         fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
 
+        save_fig_as_html_div(fig, filename)
+
         if return_as_dict:
             return fig.to_dict()
         else:
@@ -1204,7 +1334,7 @@ def plotly_barchart_weekdays(
     except ValueError as e:
         return failed_chart_message_div(
             "Could not resolve chart. Try filtering the data to reduce the number of systems.",
-            e
+            e,
         )
 
 
@@ -1257,7 +1387,9 @@ def plotly_frequency_barchart(
     n_facet_rows = 1
 
     if params["facet_col"]:
-        chart_height, n_facet_rows = calc_facet_rows_and_height(df, params["facet_col"], params["facet_col_wrap"])
+        chart_height, n_facet_rows = calc_facet_rows_and_height(
+            df, params["facet_col"], params["facet_col_wrap"]
+        )
 
     n_colours = len(df_aggregated[df_legend_col].unique())
     colour_sequence = calculate_colour_sequence(params["colourmap"], n_colours)
@@ -1290,7 +1422,7 @@ def plotly_frequency_barchart(
 
     fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
 
-    csv_data_frequency(fig, params)
+    save_fig_as_html_div(fig, params["filename"])
 
     if params["return_as_dict"]:
         return fig.to_dict(), None
@@ -1300,12 +1432,17 @@ def plotly_frequency_barchart(
             csv_name,
         )
 
-        return plot(
-            fig,
-            output_type="div",
-            include_plotlyjs=False,
-            config=global_config(params["file_name"], height_multiplier=chart_height / 500.0),
-        ), csv_data
+        return (
+            plot(
+                fig,
+                output_type="div",
+                include_plotlyjs=False,
+                config=global_config(
+                    params["filename"], height_multiplier=chart_height / 500.0
+                ),
+            ),
+            csv_data,
+        )
 
 
 def construct_over_time_charts(
@@ -1332,7 +1469,7 @@ def construct_over_time_charts(
                                   "M" (months), "W" (weeks), "D" (days)
     :param params["grouping_choice"]: (string) "series" or "system"
     :param params["colourmap"]: (string) colourmap to use
-    :param params["file_name"]: (string) default filename to use for plot bitmap export
+    :param params["filename"]: (string) default filename to use for plot bitmap export
     :param params["facet_col_wrap"]: (int) number of subplots per row
     :param params["return_as_dict"]: (boolean) flag to trigger return as a dictionary rather than a HTML DIV
     :param group_by_physician: boolean flag to set whether to group by physician name
@@ -1387,7 +1524,7 @@ def construct_over_time_charts(
         "name_axis_title": params["date_title"],
         "legend_title": params["name_title"],
         "colourmap": params["colourmap"],
-        "filename": params["file_name"],
+        "filename": params["filename"],
         "facet_col_wrap": params["facet_col_wrap"],
         "sorted_category_list": sorted_categories,
         "return_as_dict": params["return_as_dict"],
@@ -1409,7 +1546,9 @@ def construct_over_time_charts(
     return return_value
 
 
-def download_link(object_to_download, download_filename, download_link_text="Download csv"):
+def download_link(
+    object_to_download, download_filename, download_link_text="Download csv"
+):
     """
     Adapted from:
     https://discuss.streamlit.io/t/heres-a-download-function-that-works-for-dataframes-and-txt/4052